src/third_party/blink/web_tests/external/wpt/html/rendering/replaced-elements/attributes-for-embedded-content-and-images/embedded-and-images-presentational-hints-ascii-case-insensitive-expected.txt

This is a testharness.js-based test.
PASS keyword absbottom
FAIL keyword abscenter assert_equals: lowercase valid expected "middle" but got "baseline"
PASS keyword absmiddle
Harness: the test ran to completion.
